+++ src/Annotation/TypeScriptInterface.php
@@ -2,8 +2,6 @@
 
 namespace Paneon\PhpToTypeScript\Annotation;
 
-use Attribute;
-
 /**
  * Class TypeScriptInterface
  *
@@ -11,7 +9,7 @@ use Attribute;
  * @Annotation
  * @Target("CLASS")
  */
-#[Attribute(flags: Attribute::TARGET_CLASS)]
+#[\Attribute(flags: \Attribute::TARGET_CLASS)]
 class TypeScriptInterface
 {
     public $if = true;