+++ src/Attribute/TypeScriptInterface.php
@@ -2,14 +2,12 @@
 
 namespace Paneon\PhpToTypeScript\Attribute;
 
-use Attribute;
-
 /**
  * Marks a class for TypeScript interface generation.
  *
  * @deprecated Use \Paneon\PhpToTypeScript\Attribute\TypeScript instead
  */
-#[Attribute(flags: Attribute::TARGET_CLASS)]
+#[\Attribute(flags: \Attribute::TARGET_CLASS)]
 class TypeScriptInterface
 {
 }